requirements.txt
streamlit>=1.28.0
supabase>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pypdf2>=3.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
Pillow>=10.0.0
pytesseract>=0.3.10
tiktoken>=0.5.1
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
requests>=2.31.0

